requirements.txt
backcall==0.2.0
black==24.2.0
click==8.1.7
colorama==0.4.4
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
ipykernel==6.6.0
ipython==7.30.1
jedi==0.18.1
jupyter-client==7.1.0
jupyter-core==4.9.1
matplotlib-inline==0.1.3
mypy-extensions==1.0.0
nest-asyncio==1.5.4
packaging==23.2
paho-mqtt==2.0.0
parso==0.8.3
pathspec==0.12.1
pickleshare==0.7.5
platformdirs==4.2.0
prompt-toolkit==3.0.24
Pygments==2.10.0
python-dateutil==2.8.2
pywin32==303
PyYAML==6.0.1
pyzmq==22.3.0
six==1.16.0
tomli==2.0.1
tornado==6.1
typing_extensions==4.10.0
wcwidth==0.2.5